+++ src/main/java/com/bootcamp/paymentdemo/domain/order/dto/OrderCreateResponse.java
@@ -14,13 +14,13 @@ public class OrderCreateResponse {
 
     @JsonSerialize(using = ToStringSerializer.class)
     private Long orderId;
-    private Integer totalAmount;
+    private Long totalAmount;
     private String orderNumber;
 
     private OrderCreateResponse(
             Long memberId,
             Long orderId,
-            Integer totalAmount,
+            Long totalAmount,
             String orderNumber
     ) {
         this.memberId = memberId;
@@ -32,7 +32,7 @@ public class OrderCreateResponse {
     public static OrderCreateResponse register(
             Long memberId,
             Long orderId,
-            Integer totalAmount,
+            Long totalAmount,
             String orderNumber
     ) {
         return new OrderCreateResponse(memberId, orderId, totalAmount, orderNumber);
